ai/app/requirements.txt
flask
flask_cors
requests
python-dotenv
azure-cognitiveservices-vision-customvision